requirements.txt
discord.py[voice]
aiohttp
asyncio
pillow
captcha
pytz
requests